pyproject.toml
[tool.poetry]
name = "c-py-py"
version = "0.1.5"
description = ""
authors = ["torshepherd <tor.aksel.shepherd@gmail.com>"]
license = "MIT"
readme = "README.md"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include = "cpp", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
invoke-iife = "^1.1.0"
typing-extensions = "^4.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
